+++ src/components/views/LocalModList.vue
@@ -52,10 +52,10 @@
             </template>
             <template v-slot:body>
                 <div v-if="dependencyListDisplayType === 'disable'" class='notification is-warning'>
-                    <p>Other mods depend on this mod. Disabling this mod will disable all other dependants.</p>
+                    <p>Other mods depend on this mod. Select <strong>Disable all</strong> to disable dependent mods which may cause errors.</p>
                 </div>
                 <div v-if="dependencyListDisplayType === 'uninstall'" class='notification is-warning'>
-                    <p>Other mods depend on this mod. Uninstalling this mod will remove all mods that depend on it.</p>
+                    <p>Other mods depend on this mod. Select <strong>Uninstall all</strong> to uninstall dependent mods which may cause errors.</p>
                 </div>
                 <p v-if="dependencyListDisplayType === 'disable'">Mods to be disabled:</p>
                 <p v-if="dependencyListDisplayType === 'uninstall'">Mods to be uninstalled:</p>
@@ -93,13 +93,26 @@
             </template>
             <template v-slot:footer>
                 <button v-if="dependencyListDisplayType === 'disable'" class="button is-info"
-                        @click="disableMod(selectedManifestMod)">
-                    Disable
+                        @click="disableModWithDependents(selectedManifestMod)">
+                    Disable all (recommended)
+                </button>
+                <button v-if="dependencyListDisplayType === 'disable'" class="button"
+                        @click="disableModExcludeDependents(selectedManifestMod)">
+                    Disable {{selectedManifestMod.getName()}} only
                 </button>
                 <button v-if="dependencyListDisplayType === 'uninstall'" class="button is-info"
-                        @click="uninstallMod(selectedManifestMod)">
-                    Uninstall
+                        :disabled="modBeingUninstalled !== null"
+                        @click="uninstallModWithDependents(selectedManifestMod)">
+                    Uninstall all (recommended)
                 </button>
+                <button v-if="dependencyListDisplayType === 'uninstall'" class="button"
+                        :disabled="modBeingUninstalled !== null"
+                        @click="uninstallModExcludeDependents(selectedManifestMod)">
+                    Uninstall {{selectedManifestMod.getName()}} only
+                </button>
+                <span v-if="modBeingUninstalled" class="tag is-warning margin-top--1rem margin-left">
+                    Uninstalling {{ modBeingUninstalled }}
+                </span>
                 <button v-if="dependencyListDisplayType === 'view'" class="button is-info"
                         @click="selectedManifestMod = null">
                     Done
@@ -138,12 +151,15 @@
                             Disabled
                         </span>
                         <span class="card-title selectable">
-                            <template v-if="key.isEnabled()">
-                                {{key.getDisplayName()}} <span class="card-byline selectable">by {{key.getAuthorName()}}</span>
-                            </template>
-                            <template v-else>
-                                <strike class='selectable'>{{key.getDisplayName()}} <span class="card-byline">by {{key.getAuthorName()}}</span></strike>
-                            </template>
+                            <component :is="key.isEnabled() ? 'span' : 'strike'" class="selectable">
+                                {{key.getDisplayName()}}
+                                <span class="selectable card-byline">
+                                    v{{key.getVersionNumber()}}
+                                </span>
+                                <span :class="`card-byline ${key.isEnabled() && 'selectable'}`">
+                                    by {{key.getAuthorName()}}
+                                </span>
+                            </component>
                         </span>
                     </span>
                 </template>
@@ -160,8 +176,8 @@
                         <i class='fas fa-cloud-upload-alt' v-tooltip.left="'An update is available'"></i>
                     </span>
                     <span class='card-header-icon'
-                          v-if="getMissingDependencies(key).length > 0">
-                        <i class='fas fa-exclamation-circle' v-tooltip.left="`Missing ${getMissingDependencies(key).length} dependencies`"></i>
+                          v-if="getDisabledDependencies(key).length > 0 || getMissingDependencies(key).length > 0">
+                        <i class='fas fa-exclamation-circle' v-tooltip.left="`There is an issue with the dependencies for this mod`"></i>
                     </span>
                     <span class='card-header-icon'
                           @click.prevent.stop="() => key.isEnabled() ? disableModRequireConfirmation(key) : enableMod(key)">
@@ -179,16 +195,20 @@
                     <a class='card-footer-item' @click="enableMod(key)" v-else>Enable</a>
                 </template>
                 <a class='card-footer-item' @click="viewDependencyList(key)">Associated</a>
-                <Link :url="`${key.getWebsiteUrl()}${key.getVersionNumber().toString()}`"
+                <Link :url="key.getWebsiteUrl()"
                       :target="'external'"
                       class="card-footer-item">
-                        <i class='fas fa-code-branch margin-right margin-right--half-width'></i>
-                        {{key.getVersionNumber().toString()}}
+                        Website
+                        <i class="fas fa-external-link-alt margin-left margin-left--half-width"></i>
                 </Link>
                 <a class='card-footer-item' v-if="!isLatest(key)" @click="updateMod(key)">Update</a>
                 <a class='card-footer-item' v-if="getMissingDependencies(key).length > 0"
                    @click="downloadDependency(getMissingDependencies(key)[0])">
                     Download dependency
+                </a>
+                <a class='card-footer-item' v-if="getDisabledDependencies(key).length > 0"
+                   @click="enableMod(getDisabledDependencies(key)[0])">
+                    Enable {{getDisabledDependencies(key)[0].getDisplayName()}}
                 </a>
                 <template v-if="getThunderstoreModFromMod(key) !== undefined">
                     <template v-if="getThunderstoreModFromMod(key).getDonationLink() !== undefined">
@@ -261,6 +281,7 @@ import SearchUtils from '../../utils/SearchUtils';
         private showingDependencyList: boolean = false;
         private selectedManifestMod: ManifestV2 | null = null;
         private dependencyListDisplayType: string = 'view';
+        private modBeingUninstalled: string | null = null;
 
         // Filtering
         private sortDisabledPosition: SortLocalDisabledMods = this.settings.getInstalledDisablePosition();
@@ -324,7 +345,18 @@ import SearchUtils from '../../utils/SearchUtils';
         }
 
         getThunderstoreModFromMod(mod: ManifestV2) {
-            return ModBridge.getThunderstoreModFromMod(mod, this.thunderstorePackages);
+            return ModBridge.getCachedThunderstoreModFromMod(mod);
+        }
+
+        async updateModListAfterChange(updatedList: ManifestV2[]) {
+            await this.$store.dispatch("updateModList", updatedList);
+
+            const err = await ConflictManagementProvider.instance.resolveConflicts(updatedList, this.contextProfile!);
+            if (err instanceof R2Error) {
+                this.$emit('error', err);
+            }
+
+            this.filterModList();
         }
 
         async moveUp(vueMod: any) {
@@ -334,12 +366,7 @@ import SearchUtils from '../../utils/SearchUtils';
                 this.$emit('error', updatedList);
                 return;
             }
-            await this.$store.dispatch("updateModList",updatedList);
-            const err = await ConflictManagementProvider.instance.resolveConflicts(updatedList, this.contextProfile!);
-            if (err instanceof R2Error) {
-                this.$emit('error', err);
-            }
-            this.filterModList();
+            await this.updateModListAfterChange(updatedList);
         }
 
         async moveDown(vueMod: any) {
@@ -349,16 +376,11 @@ import SearchUtils from '../../utils/SearchUtils';
                 this.$emit('error', updatedList);
                 return;
             }
-            await this.$store.dispatch("updateModList", updatedList);
-            const err = await ConflictManagementProvider.instance.resolveConflicts(updatedList, this.contextProfile!);
-            if (err instanceof R2Error) {
-                this.$emit('error', err);
-            }
-            this.filterModList();
+            await this.updateModListAfterChange(updatedList);
         }
 
-        isLatest(vueMod: any): boolean {
-            return ModBridge.isLatestVersion(vueMod);
+        isLatest(mod: ManifestV2): boolean {
+            return ModBridge.isCachedLatestVersion(mod);
         }
 
         getMissingDependencies(vueMod: any): string[] {
@@ -369,6 +391,21 @@ import SearchUtils from '../../utils/SearchUtils';
             });
         }
 
+        getDisabledDependencies(vueMod: any): ManifestV2[] {
+            const mod: Mod = new Mod().fromReactive(vueMod);
+            const installedMods = [...this.modifiableModList];
+
+            const installedDependencies = mod.getDependencies().filter((dependency: string) => {
+                return this.modifiableModList.find((localMod: ManifestV2) => dependency.toLowerCase().startsWith(localMod.getName().toLowerCase() + "-"));
+            })
+                .filter(value => installedMods.find(installed => value.toLowerCase().startsWith(installed.getName().toLowerCase() + "-")))
+                .map(value => installedMods.find(installed => value.toLowerCase().startsWith(installed.getName().toLowerCase() + "-")));
+
+            const safeInstalledDependencies = installedDependencies as ManifestV2[];
+
+            return safeInstalledDependencies.filter(value => !value.isEnabled());
+        }
+
         getDependantList(mod: ManifestV2): Set<ManifestV2> {
             return Dependants.getDependantList(mod, this.modifiableModList);
         }
@@ -377,7 +414,7 @@ import SearchUtils from '../../utils/SearchUtils';
             return Dependants.getDependencyList(mod, this.modifiableModList);
         }
 
-        async performUninstallMod(mod: ManifestV2): Promise<R2Error | void> {
+        async performUninstallMod(mod: ManifestV2, updateModList=true): Promise<ManifestV2[] | R2Error> {
             const uninstallError: R2Error | null = await ProfileInstallerProvider.instance.uninstallMod(mod, this.contextProfile!);
             if (uninstallError instanceof R2Error) {
                 // Uninstall failed
@@ -392,17 +429,25 @@ import SearchUtils from '../../utils/SearchUtils';
                 this.$emit('error', modList);
                 return modList;
             }
-            await this.$store.dispatch("updateModList",modList);
-            const err = await ConflictManagementProvider.instance.resolveConflicts(modList, this.contextProfile!);
-            if (err instanceof R2Error) {
-                this.$emit('error', err);
+            if (updateModList) {
+                await this.updateModListAfterChange(modList);
             }
+            return modList;
         }
 
-        async disableMod(vueMod: any) {
+        async disableModWithDependents(vueMod: any) {
             const mod: ManifestV2 = new ManifestV2().fromReactive(vueMod);
+            await this.disableMods([...Dependants.getDependantList(mod, this.modifiableModList), mod]);
+        }
+
+        async disableModExcludeDependents(vueMod: any) {
+            const mod: ManifestV2 = new ManifestV2().fromReactive(vueMod);
+            await this.disableMods([mod]);
+        }
+
+        async disableMods(modsToDisable: ManifestV2[]) {
             try {
-                const result = await this.performDisable([...Dependants.getDependantList(mod, this.modifiableModList), mod]);
+                const result = await this.performDisable(modsToDisable);
                 if (result instanceof R2Error) {
                     this.$emit('error', result);
                     return;
@@ -435,34 +480,42 @@ import SearchUtils from '../../utils/SearchUtils';
                 this.$emit('error', updatedList);
                 return updatedList;
             }
-            await this.$store.dispatch("updateModList", updatedList);
-            const err = await ConflictManagementProvider.instance.resolveConflicts(updatedList, this.contextProfile!);
-            if (err instanceof R2Error) {
-                this.$emit('error', err);
-            }
-            this.filterModList();
+            await this.updateModListAfterChange(updatedList);
         }
 
-        async uninstallMod(vueMod: any) {
+        async uninstallModWithDependents(vueMod: any) {
             let mod: ManifestV2 = new ManifestV2().fromReactive(vueMod);
+            await this.uninstallMods([...Dependants.getDependantList(mod, this.modifiableModList), mod]);
+        }
+
+        async uninstallModExcludeDependents(vueMod: any) {
+            let mod: ManifestV2 = new ManifestV2().fromReactive(vueMod);
+            await this.uninstallMods([mod]);
+        }
+
+        async uninstallMods(modsToUninstall: ManifestV2[]) {
+            let lastSuccess: ManifestV2[] | null = null;
             try {
-                for (const dependant of Dependants.getDependantList(mod, this.modifiableModList)) {
-                    const result = await this.performUninstallMod(dependant);
+                for (const mod of modsToUninstall) {
+                    const result = await this.performUninstallMod(mod, false);
                     if (result instanceof R2Error) {
                         this.$emit('error', result);
+                        this.modBeingUninstalled = null;
                         return;
+                    } else {
+                        lastSuccess = result;
                     }
-                }
-                const result = await this.performUninstallMod(mod);
-                if (result instanceof R2Error) {
-                    this.$emit('error', result);
-                    return;
                 }
             } catch (e) {
                 // Failed to uninstall mod.
                 const err: Error = e as Error;
                 this.$emit('error', err);
                 LoggerProvider.instance.Log(LogSeverity.ACTION_STOPPED, `${err.name}\n-> ${err.message}`);
+            } finally {
+                this.modBeingUninstalled = null;
+                if (lastSuccess) {
+                    await this.updateModListAfterChange(lastSuccess);
+                }
             }
             this.selectedManifestMod = null;
             const result: ManifestV2[] | R2Error = await ProfileModList.getModList(this.contextProfile!);
@@ -470,26 +523,19 @@ import SearchUtils from '../../utils/SearchUtils';
                 this.$emit('error', result);
                 return;
             }
-            await this.$store.dispatch("updateModList", result);
-            const err = await ConflictManagementProvider.instance.resolveConflicts(result, this.contextProfile!);
-            if (err instanceof R2Error) {
-                this.$emit('error', err);
-            }
-            this.filterModList();
+            await this.updateModListAfterChange(result);
         }
 
         showDependencyList(vueMod: any, displayType: string) {
             this.selectedManifestMod = new ManifestV2().fromReactive(vueMod);
             this.dependencyListDisplayType = displayType;
             this.showingDependencyList = true;
-            console.log(this.dependencyListDisplayType, this.showingDependencyList)
         }
 
         uninstallModRequireConfirmation(vueMod: any) {
             const mod: ManifestV2 = new ManifestV2().fromReactive(vueMod);
             if (this.getDependantList(mod).size === 0) {
                 this.performUninstallMod(mod);
-                this.filterModList();
             } else {
                 this.showDependencyList(mod, DependencyListDisplayType.UNINSTALL);
             }
@@ -497,17 +543,13 @@ import SearchUtils from '../../utils/SearchUtils';
 
         disableModRequireConfirmation(vueMod: any) {
             const mod: ManifestV2 = new ManifestV2().fromReactive(vueMod);
-            const enabledDependants: ManifestV2[] = [];
-            this.getDependantList(mod).forEach(value => {
+            for (const value of this.getDependantList(mod)) {
                if (value.isEnabled()) {
-                   enabledDependants.push(value);
+                   this.showDependencyList(mod, DependencyListDisplayType.DISABLE);
+                   return;
                }
-            });
-            if (enabledDependants.length === 0) {
-                this.performDisable([mod]);
-            } else {
-                this.showDependencyList(mod, DependencyListDisplayType.DISABLE);
             }
+            this.performDisable([mod]);
         }
 
         viewDependencyList(vueMod: any) {
@@ -549,12 +591,7 @@ import SearchUtils from '../../utils/SearchUtils';
                 this.$emit('error', updatedList);
                 return updatedList;
             }
-            await this.$store.dispatch("updateModList",updatedList);
-            const err = await ConflictManagementProvider.instance.resolveConflicts(updatedList, this.contextProfile!);
-            if (err instanceof R2Error) {
-                this.$emit('error', err);
-            }
-            this.filterModList();
+            await this.updateModListAfterChange(updatedList);
         }
 
         updateMod(vueMod: any) {