+++ src/r2mm/installing/default_installation_rules/InstallationRuleApplicator.ts
@@ -113,6 +113,13 @@ export default class InstallationRuleApplicator {
             buildBepInExRules("Sunkenland"),
             buildBepInExRules("Atomicrops"),
             buildBepInExRules("Erenshor"),
+            buildBepInExRules("LastTrainOuttaWormtown"),
+            buildBepInExRules("Dredge"),
+            buildBepInExRules("CitiesSkylines2"),
+            buildBepInExRules("LethalCompany"),
+            buildBepInExRules("MeepleStation"),
+            buildBepInExRules("VoidCrew"),
+            buildBepInExRules("Sailwind"),
             InstallRules_VotV(),
         ]
     }